Commint all headers and src into new folder
This folder contains all files that need to be merged into main code,
However the music part still not compiling on my machine but that is due
to the .bat not set to my paths. Also .exe is in images folder for just
a basic show of what is working).